requirements.txt
fastapi
uvicorn
paramiko
pyyaml
requests
python-dotenv
